Fix login
At some point there was a quiz added to the login page. The quiz is rendered using javascript, why the use of `requests_html`[1] here.

I did the least amount of changes needed in order to be able to login. Long-term one should probably use `requests_html` for everything instead of mixing usage with "pure" `requests`.

[1]: https://github.com/psf/requests-html